BrotherRAT/requirements.txt
flask>=3.0
flask-sock>=0.7
customtkinter>=5.2
Pillow>=10.0